requirements.txt
# Docker管理相关
docker==6.1.3

# Kubernetes管理相关
kubernetes==27.2.0

# Web API框架
fastapi==0.104.1
uvicorn==0.24.0

# 配置管理
PyYAML==6.0.1

# 命令行界面
rich==13.7.0
click==8.1.7

# 日志处理
loguru==0.7.2

# 数据验证
pydantic==2.5.0

# HTTP客户端
requests==2.31.0

# 异步支持
aiofiles==23.2.1

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
black==23.11.0
flake8==6.1.0